+++ NewProject/Areas/Admin/Controllers/ProductsController.cs
@@ -15,80 +15,154 @@ namespace NewProject.Areas.Admin.Controllers
     public class ProductsController : Controller
     {
         private Model1 db = new Model1();
+        public bool Phanquyen()
+        {
+            var session = (LoginModels)Session[LoginConstants.LOGIN_SESSION];
+            if (session == null)
+            {
+
+                return false;
+
+            }
+            else
+            {
+                var account = new AccountDao();
+                var id = account.Byname(session.username);
+                if (id != 2)
+                {
+                    return false;
+
+                }
+            }
+            return true;
+
+
+        }
 
         // GET: Admin/Products
         public ActionResult Index(string Searchstring, int page = 1, int pagesize = 20)
         {
-            var ac = new AccountDao();
-            var session = (LoginModels)Session[LoginConstants.LOGIN_SESSION];
-            if (session == null)
+            if (Phanquyen())
             {
-                return RedirectToAction("Error", "Home1");
+                var product = new ProductsDao();
+                var listproduct = product.DanhSachSanPham(Searchstring, page, pagesize);
+                ViewBag.Search = Searchstring;
+                return View(listproduct);
             }
             else
             {
-                var account = ac.Byname(session.username);
-                if (account != 2)
-                {
-                    return RedirectToAction("Error", "Home1");
-                }
-                else
-                {
-                    var pr = new ProductsDao();
-                    var list = pr.Pagelist_sa(Searchstring, page, pagesize);
-                    ViewBag.Search = Searchstring;
-                    return View(list);
-                }
+                return RedirectToAction("Error", "Home1");
+            }
+        }
+        public ActionResult ThungRacSanPham(string Searchstring, int page = 1, int pagesize = 20)
+        {
+            if (Phanquyen())
+            {
+                var product = new ProductsDao();
+                var listproduct = product.DanhSachThungRac(Searchstring, page, pagesize);
+                ViewBag.Search = Searchstring;
+                return View(listproduct);
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+          
+        }
+        public ActionResult KhoiPhucSanPham(int id)
+        {
+            if (Phanquyen())
+            {
+                var pr = new ProductsDao();
+                pr.Bo_KhoiPhuc_SanPham(id, false);
+                return RedirectToAction("ThungRacSanPham");
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+           
+        }
+        public ActionResult BoThungRac(int id)
+        {
+            if (Phanquyen())
+            {
+                var pr = new ProductsDao();
+                pr.Bo_KhoiPhuc_SanPham(id, true);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
             }
 
-
+           
         }
-        public ActionResult Statefalse(string Searchstring, int page = 1, int pagesize = 20)
+        public ActionResult XoaSanPham(int id)
         {
-            var pr = new ProductsDao();
-            var list = pr.Pagelist_false(Searchstring, page, pagesize);
-            ViewBag.Search = Searchstring;
-            return View(list);
-        }
-        public ActionResult Restore(int id)
-        {
-            var pr = new ProductsDao();
-            pr.SetStates(id, true);
-            return RedirectToAction("Statefalse");
-        }
-        public ActionResult Del(int id)
-        {
-            var pr = new ProductsDao();
-            pr.SetStates(id, false);
-            return RedirectToAction("Index");
-        }
-        public ActionResult Dele(int id)
-        {
-            var pr = new ProductsDao();
-            pr.Delete(id);
-            return RedirectToAction("Statefalse");
+            if (Phanquyen())
+            {
+                var pr = new ProductsDao();
+                pr.Delete(id);
+                return RedirectToAction("ThungRacSanPham");
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+           
         }
 
+        public ActionResult Ban_DungBan(int id)
+        {
+            if (Phanquyen())
+            {
+                var pr = new ProductsDao();
+                pr.Ban_DungBan(id);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+
+        }
         // GET: Admin/Products/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Phanquyen())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Sach sach = db.Saches.Find(id);
+                if (sach == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(sach);
             }
-            Sach sach = db.Saches.Find(id);
-            if (sach == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Error", "Home1");
             }
-            return View(sach);
+           
         }
 
         // GET: Admin/Products/Create
         public ActionResult Create()
         {
-            ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach");
-            return View();
+            if (Phanquyen())
+            {
+                ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+           
         }
 
         // POST: Admin/Products/Create
@@ -98,31 +172,48 @@ namespace NewProject.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,ID_LoaiSach,TenSach,Anh,SoLuong,SoLuong_DaBan,Mota,Gia,Gia_giam,Trangthai,NhaXuatBan,TacGia")] Sach sach)
         {
-            if (ModelState.IsValid)
+            if (Phanquyen())
             {
-                db.Saches.Add(sach);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Saches.Add(sach);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+
+                ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
+                return View(sach);
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
             }
 
-            ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
-            return View(sach);
+            
         }
 
         // GET: Admin/Products/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Phanquyen())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Sach sach = db.Saches.Find(id);
+                if (sach == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
+                return View(sach);
             }
-            Sach sach = db.Saches.Find(id);
-            if (sach == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Error", "Home1");
             }
-            ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
-            return View(sach);
+            
         }
 
         // POST: Admin/Products/Edit/5
@@ -130,31 +221,48 @@ namespace NewProject.Areas.Admin.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID,ID_LoaiSach,TenSach,Anh,SoLuong,SoLuong_DaBan,Mota,Gia,Gia_giam,Trangthai,NhaXuatBan,TacGia")] Sach sach)
+        public ActionResult Edit([Bind(Include = "ID,ID_LoaiSach,TenSach,Anh,SoLuong,SoLuong_DaBan,Mota,Gia,Gia_giam,Trangthai,NhaXuatBan,TacGia,DeleteStatus")] Sach sach)
         {
-            if (ModelState.IsValid)
+            if (Phanquyen())
             {
-                db.Entry(sach).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(sach).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
+                return View(sach);
             }
-            ViewBag.ID_LoaiSach = new SelectList(db.Loai_Sach, "ID", "Loaisach", sach.ID_LoaiSach);
-            return View(sach);
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+
+           
         }
 
         // GET: Admin/Products/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Phanquyen())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Sach sach = db.Saches.Find(id);
+                if (sach == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(sach);
             }
-            Sach sach = db.Saches.Find(id);
-            if (sach == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Error", "Home1");
             }
-            return View(sach);
+            
         }
 
         // POST: Admin/Products/Delete/5
@@ -162,10 +270,19 @@ namespace NewProject.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Sach sach = db.Saches.Find(id);
-            db.Saches.Remove(sach);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Phanquyen())
+            {
+                Sach sach = db.Saches.Find(id);
+                db.Saches.Remove(sach);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Error", "Home1");
+            }
+
+           
         }
 
         protected override void Dispose(bool disposing)
